// src/components/types.ts
export type Colors = {
    primary: {
        light: string;
        main: string;
        dark: string;
        gradient: string;
    };
    neutral: {
        50: string;
        100: string;
        200: string;
        300: string;
        400: string;
        500: string;
        600: string;
        700: string;
        800: string;
        900: string;
    };
};